fix(checker): infer nested type-parameter predicate targets (#10797)
When a user-defined type guard's type parameter appears only *nested*
inside a parameter type — a generic alias or wrapper such as
`MaybeAsync<T> = T | AsyncIterable<T>`, `Box<T>`, or `Promise<T>` —
`resolve_generic_predicate` failed to instantiate the predicate target and
left it generic (`AsyncIterable<T>`). Flow narrowing then intersected the
already-narrowed value with the generic target, producing an
over-constrained `AsyncIterable<string> & AsyncIterable<any>` instead of
`AsyncIterable<string>`.

Structural rule: when a predicate's type parameter is nested in a compound
predicate target, infer it from the argument the same way call resolution
does, then instantiate the target. Bare-`T` predicate targets keep their
existing union-subtraction path so the false branch still narrows correctly.

Owner layer: inference is a solver concern. A new solver primitive
`infer_type_arguments_from_param_args` reuses the existing `InferenceContext`
(`infer_from_types` + `fix_current_variables`) and is reached through the
`flow_analysis` query boundary; the checker only orchestrates.

Adjacent cases covered by tests (renamed binders, distinct element types,
`for await`/property-access element usage). Verified the bare-`T` false-branch
case (`value is T` over `T | { data: T }`) is unaffected.
