+++ crates/tsz-checker/Cargo.toml
@@ -1699,6 +1699,10 @@ name = "class_expression_implements_display_tests"
 path = "tests/class_expression_implements_display_tests.rs"
 
 [[test]]
+name = "predicate_alias_generic_inference_tests"
+path = "tests/predicate_alias_generic_inference_tests.rs"
+
+[[test]]
 name = "class_implements_predicate_inference_tests"
 path = "tests/class_implements_predicate_inference_tests.rs"
